Move intltool to PKG_NAMES_ALL
As a SUSE user, `scanmem` failed to compile because it did not have this dependency. I am unsure why it was omitted but moving it to `PKG_NAMES_ALL` and running `install_pince.sh` did the trick, and PINCE can now properly run. Since all distros: Debian, SUSE (now), Fedora, and Arch use the dependency, I felt it was fine to move it to the the `PKG_NAMES_ALL`