Update SysUtils.py
due to regex used to validate a line returned by ps -eo, each line that doesn't begin with blank char is ignored. This results in processes with PID higher than 999999 to be ignored and not listed in process selector. This trivial change works around the problem by forcing PID column in ps output to always be 11 characters wide.